refactor: update imports to new entity entrypoints
