+++ src/widgets/brand-products/model/useBrandProducts.js
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from 'react';
 
-import { getProductsByBrand } from '@/entities/product/api/getProductsByBrand';
+import { getProductsByBrand } from '@/entities/product/api/client';
 
 export function useBrandProducts(brand) {
   const b = String(brand || '');
